Move dumping of Frames to library
